+++ bitframe-client/ui/react/src/main/kotlin/bitframe/panel/Main.kt
@@ -12,6 +12,7 @@ import bitframe.PanelPageRoute
 import bitframe.SignInPageRoute
 import bitframe.renderers.Renderer
 import react.router.dom.Redirect
+import react.useEffectOnce
 import reakt.*
 import styled.styledDiv
 import useViewModelState
@@ -24,7 +25,9 @@ private external interface PanelProps : Props {
 
 private val Panel = fc<PanelProps> { props ->
     val controller = props.controller
-    when (val state = useViewModelState(props.scope.viewModel)) {
+    val scope = props.scope
+    useEffectOnce { scope.initPanel() }
+    when (val state = useViewModelState(scope.viewModel)) {
         is PanelState.Loading -> LoadingBox(state.message, 80)
         is PanelState.Panel -> NavigationDrawer(
             drawerState = controller,