+++ bitframe-client/viewmodels/src/commonMain/kotlin/bitframe/authentication/signin/SignInIntent.kt
@@ -3,6 +3,7 @@ package bitframe.authentication.signin
 import bitframe.authentication.spaces.Space
 
 sealed class SignInIntent {
+    object InitForm : SignInIntent()
     data class Submit(val credentials: SignInCredentials) : SignInIntent()
     data class Resolve(val space: Space) : SignInIntent()
 }