+++ bitframe-client/viewmodels/src/commonMain/kotlin/bitframe/panel/PanelIntent.kt
@@ -1,3 +1,5 @@
 package bitframe.panel
 
-sealed class PanelIntent
+sealed class PanelIntent {
+    object InitPanel : PanelIntent()
+}